requirements.txt
Pillow>=11.0.0
tqdm>=4.66.0
InquirerPy>=0.3.4
